flow/rewards: rewrite to send to s3 and remove datastore
Simplify rewards processing by considering two streams of data that are
needed for calculating earned rewards:
- Unclaimed Balances
- Claimed Rewards

Given a range of heights, we can partition these heights into hourly
heights (by using the unix hour of the block's timestamp). This is known
as a crossing height.

For each crossing height we can:
- Fetch accounts, then fetch unclaimed balances at the current height
for each account.
- Fetch claimed rewards by considering transactions from the previous height + 1 to the current height.

For the first crossing height, there is no previous height, so we fetch
only the the unclaimed balances for that height.

After we have the above we can simply write it to S3. Removing the need
for datastore.

Changes made during this rewrite include:

- Simplying FetchHeights, here we only figure out crossing heights, and
nothing else. Each crossing height is also enriched with the previous
height, and the block time for each height. This makes each crossing
height self contained with enough information to fetch unclaimed
balances, and claims from transactions.
- Simplying CalculateRewards by splitting it into three steps.
  - Fetch accounts
  - Fetch unclaimed balances for each account
  - Fetch claimed rewards
- Each of the above uses an errgroup to do a simple parallel fetch.
- None of the above requires datastore since we fetch what we need on
the fly and don't rely on any previous state.
- Removing all code that calculates earned rewards since this will be
done later in Snowflake.